Improved VideoRead adaptibility.

+++ SelfAssembly/script/VideoRead_Ultimate.m
@@ -12,15 +12,12 @@ mkdir(folder+"/"+baseName);
 frame=readFrame(vid1);% resize to exactly 1080 pixels high.
 
 gray=rgb2gray(imresize(frame,[1080,NaN]));
-% imhist(gray); % according to the histogram of the image, decide the
-% desiredRange in function, filter out the peak of dark plate.
 
 bw=imbinarize(gray);
 % automatically recognize the biggest circle.
 [centers, radii, metric] = imfindcircles(bw,[450 540],'Sensitivity',0.99,'ObjectPolarity','dark');
-figure;
-imshow(bw);
-viscircles(centers, radii,'EdgeColor','b');
+% figure;imshow(bw);
+% viscircles(centers, radii,'EdgeColor','b');
 
 %rect[xmin ymin width height] y is calculated from above
 rect=zeros(1,4);
@@ -35,43 +32,64 @@ halfSize=(uniformSize+1)/2;
 mask=((X-halfSize).^2+(Y-halfSize).^2)>(uniformSize-1)^2/4;
 
 %%
-output=processFrame(frame,uniformSize,rect,mask);
+output=processFrame(frame,uniformSize,rect,mask,[0,1],1.0,false);
 figure; imshow(output);
-% figure; imshow(imbinarize(cropped));
+figure;imhist(output); % according to the histogram of the image, decide the
+% desiredRange in function, filter out the peak of dark plate.
+desiredRange=[0.1,0.2];
 
+% check the binarize effect.
+output=processFrame(frame,uniformSize,rect,mask,desiredRange,3.0,true);
+figure; imshow(output);
 %% Output Frames
 tic;
-i=1;
-while hasFrame(vid1)
+if exist('extractionMode','var')==1 && upper(extractionMode)==upper("continuous")
+    i=1;
+    while hasFrame(vid1)
+        frame=readFrame(vid1);
+        output=processFrame(frame,uniformSize,rect,mask,desiredRange,3.0,true);
+        imwrite(output,sprintf("%s/%s/%04d.png",folder,baseName,i));
+        i=i+1;
+    end
+else
+    % default is only single frame.
+    if exist('videoTime','var')==0 % if undefined, define it to be 0.
+        videoTime=0;
+    end
+    vid1.CurrentTime=videoTime;
+    i=floor(videoTime*vid1.FrameRate);
     frame=readFrame(vid1);
-    output=processFrame(frame,uniformSize,rect,mask);
+    output=processFrame(frame,uniformSize,rect,mask,desiredRange,3.0,true);
     imwrite(output,sprintf("%s/%s/%04d.png",folder,baseName,i));
-    i=i+1;
 end
-
+    
 toc
 fprintf("video extracted in %s\\%s.\n",folder,baseName);
 
 %% function
 
-function output=processFrame(input,uniformSize,rect,mask)
+function output=processFrame(input,uniformSize,rect,mask,desiredRange,gamma,binarize)
     % Gross parameter passing... Very stupid
     frame=imresize(input,[1080,NaN]);
     % image adjustment, enhance contrast.
     gray=rgb2gray(frame);
-    desiredRange=[0.25,0.5];
     gray=imadjust(gray,desiredRange,[]);  % the adjustment need to be done later, to ensure the success of circle detection
+%     gray=imadjust(gray);
     gray=imcrop(gray,rect);
     cropped=imresize(gray,[uniformSize,uniformSize]);
     cropped(mask)=255;
     % nonlinear gamma correction is the KEY!
-    cropped=gammaCorrection(cropped,3.0);
-    cropped=imbinarize(cropped);
-    
-    % morphological process
-    se=strel('disk',1);
-    cropped=imopen(cropped,se);
-    output=bwareaopen(cropped,6);
+    cropped=gammaCorrection(cropped,gamma);
+    if binarize
+        cropped=imbinarize(cropped);
+
+        % morphological process
+        se=strel('disk',1);
+        cropped=imopen(cropped,se);
+        output=bwareaopen(cropped,6);
+    else
+        output=cropped;
+    end
 end
 
 function output=gammaCorrection(input,gamma)
